Removed old file and changed name
Added correct name to files
